refactor(dh): move logic to service
